Check installed version

+++ MyModule/Public/Install-WinGet.ps1
@@ -20,18 +20,23 @@ Function Install-WinGet {
     (New-Object System.Net.WebClient).DownloadFile($vclibs, $vclibsfile)
 
     # Dependency of Microsoft.UI.Xaml.
-    $xaml = "https://www.nuget.org/api/v2/package/Microsoft.UI.Xaml/2.7.1"
-    $xamlfile = Join-Path $(Resolve-Path ~\Downloads) "Microsoft.UI.Xaml.2.7.zip"
-    (New-Object System.Net.WebClient).DownloadFile($xaml, $xamlfile)
-    Expand-Archive -Path $xamlfile -DestinationPath ~/Downloads/Microsoft.UI.Xaml.2.7
-    Add-AppxPackage ~\Downloads\Microsoft.UI.Xaml.2.7\tools\AppX\x64\Release\Microsoft.UI.Xaml.2.7.appx
+    if (-not (Get-AppxPackage | ? name -match ui.xaml.2.7)) {
+        $xaml = "https://www.nuget.org/api/v2/package/Microsoft.UI.Xaml/2.7.1"
+        $xamlfile = Join-Path $(Resolve-Path ~\Downloads) "Microsoft.UI.Xaml.2.7.zip"
+
+        (New-Object System.Net.WebClient).DownloadFile($xaml, $xamlfile)
+        Expand-Archive -Path $xamlfile -DestinationPath ~/Downloads/Microsoft.UI.Xaml.2.7
+
+        Add-AppxPackage ~\Downloads\Microsoft.UI.Xaml.2.7\tools\AppX\x64\Release\Microsoft.UI.Xaml.2.7.appx
+
+        Remove-Item $xamlfile -Force
+    }
 
     Add-AppxPackage $vclibsfile
     Add-AppxPackage $wingetfile
 
     Remove-Item $vclibsfile -Force
     Remove-Item $wingetfile -Force
-    Remove-Item $xamlfile -Force
 
     $time = $((Get-Date).subtract($start_time).seconds)
     Write-Output "Downloaded WinGet in $time seconds"
